+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `ChatClient.clearCacheAndTemporaryFiles(context: Context)` method for clearing cache and temporary files created by the SDK. [#6090](https://github.com/GetStream/stream-chat-android/pull/6090)
 
 ### ⚠️ Changed